+++ influxdb/tests/utilities.rs
@@ -1,5 +1,5 @@
 use futures::prelude::*;
-use influxdb::{Client, Error, Query};
+use influxdb::{Client, Error, ReadQuery};
 use std::panic::{AssertUnwindSafe, UnwindSafe};
 
 #[allow(dead_code)]
@@ -28,9 +28,7 @@ where
 {
     let test_name = name.into();
     let query = format!("CREATE DATABASE {}", test_name);
-    create_client(test_name)
-        .query(&Query::raw_read_query(query))
-        .await
+    create_client(test_name).query(&ReadQuery::new(query)).await
 }
 
 #[cfg(not(tarpaulin_include))]
@@ -40,9 +38,7 @@ where
 {
     let test_name = name.into();
     let query = format!("DROP DATABASE {}", test_name);
-    create_client(test_name)
-        .query(&Query::raw_read_query(query))
-        .await
+    create_client(test_name).query(&ReadQuery::new(query)).await
 }
 
 #[cfg(not(tarpaulin_include))]